Adding routes components
