+++ package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router-dom": "6",
     "react-scripts": "5.0.1",
     "sass": "^1.88.0",
     "web-vitals": "^2.1.4"
